+++ src/main/java/lotto/domain/dto/TicketDto.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.dto;
 
 import java.util.List;
