+++ src/test/java/lotto/domain/LottoMachineTest.java
@@ -3,6 +3,8 @@ package lotto.domain;
 import java.util.ArrayList;
 import java.util.List;
 
+import lotto.domain.dto.LottoResultDto;
+import lotto.domain.dto.TicketDto;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
